pyproject.toml
[project]
name = "sam3-fursearch"
version = "1.0.0"
description = "Fursuit character recognition using SAM3 and DINOv2"
requires-python = ">=3.10,<3.13"
dependencies = [
    "torch==2.5.1",
    "torchvision==0.20.1",
    "ultralytics>=8.3.0",
    "transformers>=4.40.0",
    "faiss-cpu>=1.7.4",
    "pillow>=10.0",
    "numpy>=1.24",
    "requests>=2.31.0",
    "python-telegram-bot>=20.0",
    "python-dotenv>=1.0",
    "aiohttp>=3.13.3",
    "dotenv>=0.9.9",
    "pytest>=9.0.2",
    "tenacity>=9.1.4",
]

[project.scripts]
fursearch = "sam3_fursearch.api.cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["sam3_fursearch*"]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu124" }
torchvision = { index = "pytorch-cu124" }
